Add files via upload
In this page, you can download the code of the prototype developed in our research paper titled "An Agent-based Approach to Realize Emergent Configurations in the Internet of Things". Our contribution can be found mainly in the ec package. The code comprises also 5 experiments that can be found under the experiments package. We also provide dependencies (i.e., Jar files) needed to run the project in the dependencies.zip folder.